+++ packages/fuselage-forms/src/Field/Label/index.ts
@@ -1,4 +1,7 @@
-export { default as LabelFor } from './LabelFor';
-export { default as HiddenLabel } from './HiddenLabel';
-export { default as ReferencedLabel } from './ReferencedLabel';
-export { default as FieldLabel } from './FieldLabel';
+export { default as LabelFor, type LabelForProps } from './LabelFor';
+export { default as HiddenLabel, type HiddenLabelProps } from './HiddenLabel';
+export {
+  default as ReferencedLabel,
+  type ReferencedLabelProps,
+} from './ReferencedLabel';
+export { default as FieldLabel, type FieldLabelProps } from './FieldLabel';
